+++ apps/storybook/src/Table.stories.tsx
@@ -9,8 +9,8 @@ import type {
   Column,
   Row,
   TableInstance,
-  TableState,
-} from 'react-table';
+  TableOptions,
+} from '@itwin/itwinui-react/react-table';
 import {
   Checkbox,
   Code,
@@ -21,13 +21,10 @@ import {
   Table,
   Text,
   tableFilters,
-  TableFilterValue,
-  TableProps,
   Tooltip,
   DefaultCell,
   EditableCell,
   TablePaginator,
-  TablePaginatorRendererProps,
   ActionColumn,
   Anchor,
   SelectionColumn,
@@ -36,6 +33,10 @@ import {
   Radio,
   ProgressRadial,
   BaseFilter,
+} from '@itwin/itwinui-react';
+import type {
+  TableFilterValue,
+  TablePaginatorRendererProps,
 } from '@itwin/itwinui-react';
 import { useMemo, useState } from '@storybook/addons';
 import { action } from '@storybook/addon-actions';
@@ -47,6 +48,10 @@ import {
   SvgStatusSuccess,
   SvgStatusWarning,
 } from '@itwin/itwinui-icons-react';
+import { StoryFn } from '@storybook/react';
+
+type TableProps<T extends Record<string, unknown> = Record<string, unknown>> =
+  React.ComponentProps<typeof Table<T>>;
 
 export default {
   title: 'Core/Table',
@@ -86,7 +91,7 @@ export const Basic = () => {
       },
     ],
     [],
-  );
+  ) satisfies Column[];
 
   const data = useMemo(
     () => [
@@ -95,7 +100,7 @@ export const Basic = () => {
       { name: 'Name3', description: 'Description3' },
     ],
     [],
-  );
+  ) as Record<string, unknown>[];
 
   return <Table columns={columns} data={data} emptyTableContent='No data.' />;
 };
@@ -172,7 +177,7 @@ export const SelectableMulti = () => {
         )}`,
       )(),
     [],
-  );
+  ) satisfies NonNullable<TableProps['onSelect']>;
 
   const onRowClick = useCallback(
     (event: React.MouseEvent, row: Row) =>
@@ -245,7 +250,7 @@ export const Sortable = () => {
   const onSort = useCallback(
     (state) => action(`Sort changed. Table state: ${JSON.stringify(state)}`)(),
     [],
-  );
+  ) satisfies NonNullable<TableProps['onSort']>;
 
   const columns = useMemo(
     () => [
@@ -342,7 +347,7 @@ export const Filters = () => {
   );
 
   const columns = useMemo(
-    (): Column<TableStoryDataType>[] => [
+    () => [
       {
         id: 'index',
         Header: '#',
@@ -404,7 +409,7 @@ export const Filters = () => {
       },
     ],
     [formatDate, translatedLabels],
-  );
+  ) satisfies Column<TableStoryDataType>[];
 
   const data = useMemo(
     () => [
@@ -412,7 +417,7 @@ export const Filters = () => {
         index: 1,
         name: 'Name1',
         description: 'Description1',
-        ids: ['1'],
+        ids: [1],
         startDate: new Date('May 1, 2021'),
         endDate: '2021-05-31T21:00:00.000Z',
       },
@@ -420,7 +425,7 @@ export const Filters = () => {
         index: 2,
         name: 'Name2',
         description: 'Description2',
-        ids: ['2', '3', '4'],
+        ids: [2, 3, 4],
         startDate: new Date('May 2, 2021'),
         endDate: '2021-06-01T21:00:00.000Z',
       },
@@ -428,35 +433,28 @@ export const Filters = () => {
         index: 3,
         name: 'Name3',
         description: 'Description3',
-        ids: ['3', '4'],
+        ids: [3, 4],
         startDate: new Date('May 3, 2021'),
         endDate: '2021-06-02T21:00:00.000Z',
       },
     ],
     [],
-  );
+  ) satisfies TableStoryDataType[];
 
-  const onFilter = React.useCallback(
-    (
-      filters: TableFilterValue<TableStoryDataType>[],
-      state: TableState,
-      filteredData: Row<{ name: string; description: string }>[],
-    ) => {
-      // rowInfo is used due to JSON errors when displaying row data
-      let rowInfo = '[';
-      filteredData.forEach((row) => {
-        rowInfo += `${JSON.stringify(row.original)},`;
-      });
-      rowInfo = rowInfo.slice(0, rowInfo.length - 1);
-      rowInfo += ']';
-      action(
-        `Filter changed. Filters: ${JSON.stringify(
-          filters,
-        )}, State: ${JSON.stringify(state)}, Rows: ${rowInfo}`,
-      )();
-    },
-    [],
-  );
+  const onFilter = React.useCallback((filters, state, filteredData) => {
+    // rowInfo is used due to JSON errors when displaying row data
+    let rowInfo = '[';
+    filteredData?.forEach((row) => {
+      rowInfo += `${JSON.stringify(row.original)},`;
+    });
+    rowInfo = rowInfo.slice(0, rowInfo.length - 1);
+    rowInfo += ']';
+    action(
+      `Filter changed. Filters: ${JSON.stringify(
+        filters,
+      )}, State: ${JSON.stringify(state)}, Rows: ${rowInfo}`,
+    )();
+  }, []) satisfies NonNullable<TableProps['onFilter']>;
 
   return (
     <Table
@@ -474,9 +472,10 @@ export const GlobalFilter = () => {
     description: string;
   };
 
-  const onClickHandler = (
-    props: CellProps<{ name: string; description: string }>,
-  ) => action(props.row.original.name)();
+  const onClickHandler = React.useCallback(
+    (props: CellProps<TableStoryDataType>) => action(props.row.original.name)(),
+    [],
+  );
 
   const columns = useMemo(
     () => [
@@ -507,10 +506,10 @@ export const GlobalFilter = () => {
         },
       },
     ],
-    [],
+    [onClickHandler],
   );
 
-  const data = useMemo<TableStoryDataType[]>(
+  const data = useMemo(
     () => [
       { name: 'Name1', description: 'Description7' },
       { name: 'Name2', description: 'Description7' },
@@ -554,7 +553,7 @@ export const Expandable = () => {
         )}`,
       )(),
     [],
-  );
+  ) satisfies NonNullable<TableProps['onExpand']>;
 
   const columns = useMemo(
     () => [
@@ -615,7 +614,7 @@ export const ExpandableSubrows = () => {
         )}`,
       )(),
     [],
-  );
+  ) satisfies NonNullable<TableProps['onExpand']>;
 
   const columns = useMemo(
     () => [
@@ -826,7 +825,7 @@ export const RowInViewport = () => {
 
   const onRowInViewport = useCallback((rowData) => {
     action(`Row in view: ${JSON.stringify(rowData)}`)();
-  }, []);
+  }, []) satisfies NonNullable<TableProps['onRowInViewport']>;
 
   return (
     <>
@@ -944,6 +943,11 @@ export const DisabledRows = () => {
 };
 
 export const Loading = () => {
+  type CustomStoryDataType = {
+    name: string;
+    description: string;
+  };
+
   const columns = useMemo(
     () => [
       {
@@ -959,10 +963,10 @@ export const Loading = () => {
       },
     ],
     [],
-  );
+  ) satisfies Column<CustomStoryDataType>[];
 
   return (
-    <Table
+    <Table<CustomStoryDataType>
       columns={columns}
       data={[]}
       isLoading={true}
@@ -972,6 +976,11 @@ export const Loading = () => {
 };
 
 export const NoData = () => {
+  type CustomStoryDataType = {
+    name: string;
+    description: string;
+  };
+
   const columns = useMemo(
     () => [
       {
@@ -987,10 +996,10 @@ export const NoData = () => {
       },
     ],
     [],
-  );
+  ) satisfies Column<CustomStoryDataType>[];
 
   return (
-    <Table
+    <Table<CustomStoryDataType>
       columns={columns}
       data={[]}
       isLoading={false}
@@ -1052,7 +1061,7 @@ export const ControlledState = () => {
     subRows: DemoData[];
   };
 
-  const tableInstance = React.useRef<TableInstance>();
+  const tableInstance = React.useRef<TableInstance<DemoData>>();
   const [selectedRows, setSelectedRows] = useState<DemoData[]>([]);
   const [expandedRows, setExpandedRows] = useState<DemoData[]>([]);
 
@@ -1070,9 +1079,9 @@ export const ControlledState = () => {
       },
     ],
     [],
-  );
+  ) satisfies Column<DemoData>[];
 
-  const data: DemoData[] = useMemo(
+  const data = useMemo(
     () => [
       {
         id: '1',
@@ -1135,9 +1144,24 @@ export const ControlledState = () => {
         name: 'Row 2',
         description: 'Description 2',
         subRows: [
-          { name: 'Row 2.1', description: 'Description 2.1', subRows: [] },
-          { name: 'Row 2.2', description: 'Description 2.2', subRows: [] },
-          { name: 'Row 2.3', description: 'Description 2.3', subRows: [] },
+          {
+            id: '2.1',
+            name: 'Row 2.1',
+            description: 'Description 2.1',
+            subRows: [],
+          },
+          {
+            id: '2.2',
+            name: 'Row 2.2',
+            description: 'Description 2.2',
+            subRows: [],
+          },
+          {
+            id: '2.3',
+            name: 'Row 2.3',
+            description: 'Description 2.3',
+            subRows: [],
+          },
         ],
       },
       { id: '3', name: 'Row 3', description: 'Description 3', subRows: [] },
@@ -1180,18 +1204,28 @@ export const ControlledState = () => {
       <Table
         columns={columns}
         emptyTableContent='No data.'
-        stateReducer={useCallback((newState, action, prevState, instance) => {
-          tableInstance.current = instance;
-          return newState;
-        }, [])}
+        stateReducer={
+          useCallback((newState, action, prevState, instance) => {
+            tableInstance.current = instance;
+            return newState;
+          }, []) satisfies NonNullable<TableOptions<DemoData>['stateReducer']>
+        }
         isSelectable
-        onSelect={useCallback((selected) => {
-          setSelectedRows(selected ?? []);
-        }, [])}
-        onExpand={useCallback((expanded) => {
-          setExpandedRows(expanded);
-        }, [])}
-        getRowId={useCallback((rowData) => rowData.id, [])}
+        onSelect={
+          useCallback((selected) => {
+            setSelectedRows(selected ?? []);
+          }, []) satisfies NonNullable<TableProps<DemoData>['onSelect']>
+        }
+        onExpand={
+          useCallback((expanded) => {
+            setExpandedRows(expanded ?? []);
+          }, []) satisfies NonNullable<TableProps<DemoData>['onExpand']>
+        }
+        getRowId={
+          useCallback((rowData) => rowData.id, []) satisfies NonNullable<
+            TableOptions<DemoData>['getRowId']
+          >
+        }
         data={data}
       />
     </>
@@ -1378,7 +1412,7 @@ export const Full2 = () => {
   }, []);
 
   const columns = useMemo(
-    (): Column<TableStoryDataType>[] => [
+    () => [
       {
         id: 'product',
         Header: 'Product',
@@ -1460,11 +1494,11 @@ export const Full2 = () => {
       },
     ],
     [isRowDisabled, menuItems],
-  );
+  ) satisfies Column<TableStoryDataType>[];
 
-  const rowProps = useCallback((row: Row<{ status: string | undefined }>) => {
+  const rowProps = useCallback((row: Row<TableStoryDataType>) => {
     return {
-      status: row.original.status,
+      status: row.original.status satisfies TableStoryDataType['status'],
     };
   }, []);
 
@@ -1590,7 +1624,7 @@ export const Localized = () => {
 };
 
 Localized.decorators = [
-  (Story) => (
+  (Story: StoryFn) => (
     <div style={{ height: '90vh' }}>
       <Story />
     </div>
@@ -1614,7 +1648,7 @@ export const Condensed = () => {
         )}`,
       )(),
     [],
-  );
+  ) satisfies NonNullable<TableProps['onExpand']>;
 
   const columns = useMemo(
     () => [
@@ -1754,7 +1788,7 @@ export const Editable = () => {
   );
 
   const columns = React.useMemo(
-    (): Column<TableStoryDataType>[] => [
+    () => [
       {
         id: 'name',
         Header: 'Name',
@@ -1864,7 +1898,7 @@ export const WithPaginator = () => {
 };
 
 WithPaginator.decorators = [
-  (Story) => (
+  (Story: StoryFn) => (
     <div style={{ height: '90vh' }}>
       <Story />
     </div>
@@ -2073,7 +2107,7 @@ export const WithManualPaginatorAndFilter = () => {
 };
 
 WithManualPaginatorAndFilter.decorators = [
-  (Story) => (
+  (Story: StoryFn) => (
     <div style={{ height: '90vh' }}>
       <Story />
     </div>
@@ -2212,7 +2246,7 @@ export const CustomFilter = () => {
       accessor: 'description',
       maxWidth: 200,
     },
-  ];
+  ] satisfies Column<RowData>[];
 
   return (
     <>
@@ -2231,7 +2265,7 @@ export const CustomFilter = () => {
 };
 
 CustomFilter.decorators = [
-  (Story) => (
+  (Story: StoryFn) => (
     <div style={{ height: '90vh' }}>
       <Story />
     </div>
@@ -2253,7 +2287,7 @@ export const ResizableColumns = () => {
   };
 
   const columns = useMemo(
-    (): Column<TableStoryDataType>[] => [
+    () => [
       {
         id: 'index',
         Header: '#',
@@ -2302,7 +2336,7 @@ export const ResizableColumns = () => {
       },
     ],
     [],
-  );
+  ) satisfies Column<TableStoryDataType>[];
 
   const data = useMemo(
     () => [
@@ -2335,7 +2369,7 @@ export const ResizableColumns = () => {
   );
 
   const [columnResizeMode, setColumnResizeMode] =
-    React.useState<TableProps['columnResizeMode']>('fit');
+    React.useState<TableProps<TableStoryDataType>['columnResizeMode']>('fit');
 
   return (
     <>
@@ -2511,7 +2545,7 @@ export const HorizontalScroll = () => {
   );
 
   const columns = useMemo(
-    (): Column<(typeof data)[number]>[] => [
+    () => [
       {
         id: 'product',
         Header: 'Product',
@@ -2550,7 +2584,7 @@ export const HorizontalScroll = () => {
       },
     ],
     [],
-  );
+  ) satisfies Column<(typeof data)[0]>[];
 
   return (
     <Table
@@ -2563,7 +2597,7 @@ export const HorizontalScroll = () => {
 };
 
 HorizontalScroll.decorators = [
-  (Story) => (
+  (Story: StoryFn) => (
     <div
       style={{
         height: '375px',
@@ -2675,7 +2709,7 @@ export const ScrollToRow = () => {
     [onClickHandler],
   );
 
-  const data: TableStoryDataType[] = useMemo(() => {
+  const data = useMemo(() => {
     const size = 100000;
     const arr = new Array(size);
     for (let i = 0; i < size; ++i) {
@@ -2904,7 +2938,7 @@ export const CustomizedColumns = () => {
         )}`,
       )(),
     [],
-  );
+  ) satisfies NonNullable<TableProps<(typeof data)[number]>['onExpand']>;
 
   const data = useMemo(
     () => [
@@ -2942,7 +2976,7 @@ export const CustomizedColumns = () => {
   );
 
   const columns = useMemo(
-    (): Column<(typeof data)[number]>[] => [
+    () => [
       SelectionColumn({
         isDisabled: isCheckboxDisabled,
       }),
@@ -2952,9 +2986,9 @@ export const CustomizedColumns = () => {
         Header: 'Name',
         accessor: 'name',
         cellRenderer: (props) => (
-          <DefaultCell
+          <DefaultCell<(typeof data)[number]>
             {...props}
-            isDisabled={(rowData) =>
+            isDisabled={(rowData: (typeof data)[number]) =>
               isCellDisabled(rowData) || isRowDisabled(rowData)
             }
           />
@@ -2974,7 +3008,7 @@ export const CustomizedColumns = () => {
       isCellDisabled,
       isRowDisabled,
     ],
-  );
+  ) satisfies Column<(typeof data)[number]>[];
 
   return (
     <Table
@@ -3000,10 +3034,15 @@ export const ColumnManager = () => {
     id: string;
     startDate: Date;
     endDate: Date;
+    price: string;
+    color: string;
+    stock: number;
+    rating: string;
+    location: string;
   };
 
   const columns = useMemo(
-    (): Column<TableStoryDataType>[] => [
+    () => [
       {
         id: 'index',
         Header: '#',
@@ -3074,7 +3113,7 @@ export const ColumnManager = () => {
       }),
     ],
     [],
-  );
+  ) satisfies Column<TableStoryDataType>[];
   const data = useMemo(
     () => [
       {
@@ -3245,7 +3284,7 @@ export const StickyColumns = () => {
   }, []);
 
   const columns = useMemo(
-    (): Column<(typeof data)[number]>[] => [
+    () => [
       {
         id: 'product',
         Header: 'Product',
@@ -3300,7 +3339,7 @@ export const StickyColumns = () => {
       },
     ],
     [menuItems],
-  );
+  ) satisfies Column<(typeof data)[number]>[];
 
   return (
     <Table
@@ -3314,7 +3353,7 @@ export const StickyColumns = () => {
 };
 
 StickyColumns.decorators = [
-  (Story) => (
+  (Story: StoryFn) => (
     <div
       style={{
         height: '375px',
@@ -3328,18 +3367,23 @@ StickyColumns.decorators = [
 ];
 
 export const StatusAndCellIcons = () => {
+  type CustomStoryDataType = {
+    name: string;
+    modified: string;
+    size: string;
+    startIcon?: JSX.Element;
+    endIcon?: JSX.Element;
+    status?: 'positive' | 'negative' | 'warning';
+    isLoading?: boolean;
+  };
+
   const columns = useMemo(
     () => [
       {
         id: 'name',
         Header: 'Name',
         accessor: 'name',
-        cellRenderer: (
-          props: CellRendererProps<{
-            startIcon: JSX.Element;
-            endIcon: JSX.Element;
-          }>,
-        ) => (
+        cellRenderer: (props: CellRendererProps<CustomStoryDataType>) => (
           <DefaultCell
             {...props}
             startIcon={props.cellProps.row.original.startIcon}
@@ -3358,11 +3402,7 @@ export const StatusAndCellIcons = () => {
         Header: 'Modified',
         accessor: 'modified',
         maxWidth: 200,
-        cellRenderer: (
-          props: CellRendererProps<{
-            status: 'positive' | 'warning' | 'negative' | undefined;
-          }>,
-        ) => {
+        cellRenderer: (props: CellRendererProps<CustomStoryDataType>) => {
           return (
             <DefaultCell
               {...props}
@@ -3379,7 +3419,7 @@ export const StatusAndCellIcons = () => {
       },
     ],
     [],
-  );
+  ) satisfies Column<CustomStoryDataType>[];
 
   const data = useMemo(
     () => [
@@ -3421,27 +3461,14 @@ export const StatusAndCellIcons = () => {
       },
     ],
     [],
-  );
+  ) satisfies CustomStoryDataType[];
 
-  const rowProps = useCallback(
-    (
-      row: Row<{
-        name: string;
-        modified: string;
-        size: string;
-        startIcon: JSX.Element;
-        endIcon: JSX.Element;
-        status?: 'positive' | 'negative' | 'warning';
-        isLoading?: boolean;
-      }>,
-    ) => {
-      return {
-        status: row.original.status,
-        isLoading: row.original.isLoading,
-      };
-    },
-    [],
-  );
+  const rowProps = useCallback((row: Row<CustomStoryDataType>) => {
+    return {
+      status: row.original.status,
+      isLoading: row.original.isLoading,
+    };
+  }, []);
 
   return (
     <Table