+++ packages/itwinui-react/src/core/DatePicker/DatePicker.test.tsx
@@ -548,3 +548,12 @@ it('should prevent selecting disabled dates', async () => {
   await userEvent.click(getByText('24'));
   expect(onClick).not.toHaveBeenCalled();
 });
+
+it.each([true, false])('should respect `applyBackground`', (value) => {
+  const { container } = render(<DatePicker applyBackground={value} />);
+  if (value) {
+    expect(container.querySelector('div')).toHaveClass('iui-surface');
+  } else {
+    expect(container.querySelector('div')).not.toHaveClass('iui-surface');
+  }
+});