+++ packages/itwinui-react/src/core/Table/filters/FilterToggle.tsx
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 import * as React from 'react';
 import cx from 'classnames';
-import type { HeaderGroup } from 'react-table';
+import type { HeaderGroup } from '../../../react-table/react-table.js';
 import {
   useGlobals,
   Popover,
@@ -54,14 +54,16 @@ export const FilterToggle = <T extends Record<string, unknown>>(
           content={column.render('Filter', { close, setFilter, clearFilter })}
           placement='bottom-start'
           visible={isVisible}
-          onClickOutside={close}
+          onVisibleChange={setIsVisible}
+          closeOnOutsideClick
+          applyBackground
         >
           <IconButton
             styleType='borderless'
             isActive={isVisible || isColumnFiltered}
             className={cx('iui-table-filter-button', className)}
+            aria-label='Filter'
             onClick={(e) => {
-              setIsVisible((v) => !v);
               // Prevents from triggering sort
               e.stopPropagation();
             }}