+++ packages/itwinui-react/src/core/Table/filters/types.ts
@@ -4,10 +4,9 @@
  *--------------------------------------------------------------------------------------------*/
 import type {
   FieldType,
-  FilterProps,
   FilterType,
-  HeaderGroup,
-} from 'react-table';
+  TableFilterProps,
+} from '../../../react-table/react-table.js';
 
 export interface TableFilterValue<T extends Record<string, unknown>> {
   /**
@@ -28,22 +27,4 @@ export interface TableFilterValue<T extends Record<string, unknown>> {
   filterType: FilterType<T>;
 }
 
-export type TableFilterProps<T extends Record<string, unknown> | object> =
-  FilterProps<T> & {
-    /**
-     * Data of column on which filter is opened. It is provided by the table it self.
-     */
-    column: HeaderGroup<T>;
-    /**
-     * Function to close the filter. It is provided by the table it self.
-     */
-    close: () => void;
-    /**
-     * Function to set the filter value. It is provided by the table it self.
-     */
-    setFilter: (filterValue: unknown | undefined) => void;
-    /**
-     * Function to clear the filter value. It is provided by the table it self.
-     */
-    clearFilter: () => void;
-  };
+export type { TableFilterProps };