+++ packages/itwinui-react/src/core/Table/hooks/useSelectionCell.tsx
@@ -2,7 +2,10 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import type { ColumnInstance, Hooks } from 'react-table';
+import type {
+  ColumnInstance,
+  Hooks,
+} from '../../../react-table/react-table.js';
 import { SelectionColumn, SELECTION_CELL_ID } from '../columns/index.js';
 
 export const useSelectionCell =