+++ packages/itwinui-react/src/index.ts
@@ -3,4 +3,3 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 export * from './core/index.js';
-export * from './types/react-table-config.js';