+++ packages/itwinui-react/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 3.0.0-dev.10
+
+### Major Changes
+
+- [#1514](https://github.com/iTwin/iTwinUI/pull/1514): Table now has better type support. Users must now import types from `@itwin/itwinui-react/react-table` instead of from `@types/react-table`.
+- [#1433](https://github.com/iTwin/iTwinUI/pull/1433): Updated `exports` to prevent importing internal utilities.
+
+### Minor Changes
+
+- [#1573](https://github.com/iTwin/iTwinUI/pull/1573): Added `applyBackground` prop to ColorPicker and DatePicker.
+- [#1567](https://github.com/iTwin/iTwinUI/pull/1567): Added `endIconProps` prop to ComboBox.
+
+## 3.0.0-dev.9
+
+### Major Changes
+
+- [#1506](https://github.com/iTwin/iTwinUI/pull/1506): Replaced `tippy.js` with `floating-ui` in all popover-based components. While the basic usage is unchanged, all advanced props from tippy are no longer available.
+
+  Components affected: Select, ComboBox, DropdownMenu, DropdownButton, SplitButton.
+
+### Minor Changes
+
+- [#1506](https://github.com/iTwin/iTwinUI/pull/1506): Added new `Popover` component for public use.
+
 ## 3.0.0-dev.8
 
 ### Major Changes
@@ -177,6 +201,15 @@
 
 - Updated dependencies:
   - @itwin/itwinui-css@2.0.0-dev.0
+
+## 2.12.8
+
+### Patch Changes
+
+- [#1576](https://github.com/iTwin/iTwinUI/pull/1576): Dropdown menus will now automatically become scrollable if there are too many menu items. Also added delay to submenu.
+- [#1578](https://github.com/iTwin/iTwinUI/pull/1578): Fixed missing animation in ComboBox.
+- [#1572](https://github.com/iTwin/iTwinUI/pull/1572): Adjusted borders and shadow values to be more prominent in dark theme.
+- [#1575](https://github.com/iTwin/iTwinUI/pull/1575): ProgressRadial now stays inside container when using size="x-small".
 
 ## 2.12.7